api/pyproject.toml
[project]
name = "example"
version = "0.1.0"
description = "A simple example of a FastAPI project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "psycopg[binary,pool]>=3.2.3",
    "pydantic-settings>=2.7.0",
]

[dependency-groups]
dev = ["ruff>=0.8.4"]

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "W",
    "E",   # pycodestyle
    "C90", # mccabe
    "I",   # isort
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "YTT", # flake8-2020
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "T10", # flake8-debugger
    "EM",  # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "G",   # flake8-logging-format
    "T20", # flake8-print
    "Q",   # flake8-quotes
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TID", # flake8-tidy-imports
    "DTZ", # flake8-datetimez
    "ARG", # flake8-unused-arguments
    "PGH", # pygrep-hooks
    "PLC",
    "PLE",
    "PLR",
    "PLW", # Pylint
    "RUF", # Ruff-specific rules
]
ignore = [
    "A002",
    "B008",
    "D100",
    "D106",
    "D107",
    "D203",
    "D213",
    "D406",
    "D407",
    "DTZ003",
    "FBT001",
    "FBT003",
    "ISC001",
    "N815",
    "PGH003",
    "S101",
    "T201",
]
